tic: build errors with clang
/code/illumos-gate/usr/src/cmd/tic/tic_parse.c:721:54: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
            fwrite(cBooleans, sizeof (char), BoolCount, fp) != BoolCount)
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~
/code/illumos-gate/usr/src/cmd/tic/tic_parse.c:720:51: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'short' [-Werror,-Wsign-compare]
            fwrite(namelist, sizeof (char), namelen, fp) != namelen ||
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~
/code/illumos-gate/usr/src/cmd/tic/tic_parse.c:739:8: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                                                        != l_next_free)
                                                        ^  ~~~~~~~~~~~
/code/illumos-gate/usr/src/cmd/tic/tic_parse.c:737:8: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                                                        != StrCount ||
                                                        ^  ~~~~~~~~
/code/illumos-gate/usr/src/cmd/tic/tic_parse.c:735:60: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        if (fwrite((char *)Numbers, sizeof (short), NumCount, fp) != NumCount ||
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~
5 errors generated.

/code/illumos-gate/usr/src/cmd/tic/tic_main.c:216:14: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
        if (dirnames[dirletter] == 0) {
                    ^~~~~~~~~~
/code/illumos-gate/usr/src/cmd/tic/tic_main.c:221:11: error: array subscript is of type 'char' [-Werror,-Wchar-subscripts]
                dirnames[dirletter] = 1;
                        ^~~~~~~~~~
2 errors generated.

+++ usr/src/cmd/tic/tic_main.c
@@ -213,12 +213,12 @@ check_dir(char dirletter)
 	static char	dirnames[128];
 	static char	dir[2] = " ";
 
-	if (dirnames[dirletter] == 0) {
+	if (dirnames[(int)dirletter] == 0) {
 	    dir[0] = dirletter;
 	    if (stat64(dir, &statbuf) < 0) {
 		if (mkdir(dir, 0755) < 0)
 			syserr_abort("mkdir %s returned bad status", dir);
-		dirnames[dirletter] = 1;
+		dirnames[(int)dirletter] = 1;
 	    } else if (access(dir, 7) < 0) {
 			fprintf(stderr, "%s: %s/%s: Permission denied\n",
 			    progname, destination, dir);
